backend/pyproject.toml
[project]
name = "speech-to-text-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "celery>=5.4.0",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "openai-whisper>=20240930",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]

[tool.uv]
no-build-isolation-package = ["openai-whisper"]

[tool.uv.extra-build-dependencies]
openai-whisper = ["setuptools", "pkg_resources"]